what was actually on the server in December 2022
It looks like I had a half-finished change on my desktop, using the
latest version of the elm graph-app which I was updating for use with
Numbas, and a slightly older version on the web, using the old version
of the component.

I don't want to spend time today making it work, so I've fetched the
files that were on the web.